app/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
github3.py==4.0.2
google-generativeai==0.4.0
requests==2.31.0
python-dotenv==1.0.1
